fix: add webpack
