+++ .gitignore
@@ -3,3 +3,5 @@ package-lock.json
 .stylelintrc
 .linthtml
 package.json
+package-lock.json
+webpack.config.js
