+++ app.js
@@ -1,11 +1,1 @@
-let topBtn = document.querySelector('.return_button')
-
-topBtn.onclick = function () {
-	window.scrollTo(0,0)
-}
-
-document.addEventListener('scroll', e => {
-  if (window.pageYOffset > 200) {
-    topBtn.style.opacity = '1'
-  } else { topBtn.style.opacity = '0' }
-})
+(()=>{"use strict";!function(){let t=document.querySelector(".return_button");t.onclick=function(){window.scrollTo(0,0)},document.addEventListener("scroll",(e=>{window.pageYOffset>200?t.style.opacity="1":t.style.opacity="0"}))}()})();
