+++ html/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Front-end Developer CV</title>
     <link rel="stylesheet" href="./style.css">
-<script type="module" src="app.js"></script><link href="style.css" rel="stylesheet"></head>
+</head>
 <body>
     <script src="app.js" defer></script>
     <header class="cv-header flex-row" id="main">
