+++ style.css
@@ -1,0 +1,1 @@
+@font-face{font-family:"Roboto";src:url(9d43b86421e004e704b9.woff2) format("woff2") url(d255c3ec15fbb3290d50.woff) format("woff");font-weight:400;font-style:normal;font-display:swap}@font-face{font-family:"Roboto";src:url(b21da07c934a28b2fef3.woff2) format("woff2") url(00f54dcb163757355b2d.woff) format("woff");font-weight:700;font-style:normal;font-display:swap}*,*::before,*::after{box-sizing:border-box}html{overflow-x:hidden}html,body{height:100%;font-size:62.5%;font-family:"Roboto",sans-serif;font-weight:normal;color:#000;background:#fff;scroll-behavior:smooth}img{width:100%;height:100%;object-fit:contain;flex-shrink:0}ul{padding-left:0}ul>li{list-style:none}a{text-decoration:none;color:inherit}h1,h2,h3,h4,h5,h6{padding:0}button,input{outline:none;border:none;cursor:pointer}.flex-row{display:flex;flex-direction:row;flex-wrap:wrap}header,main,footer{position:relative;width:90%;margin:0 auto}a:hover{color:rgba(185,20,20,.8)}.cv-header{margin-top:4.8rem;justify-content:space-around;align-items:center}.cv-header__title{font:normal 700 4.8rem/1.17 sans-serif}.cv-header__desc{font:normal 400 3.6rem/1.56 sans-serif}@media(max-width: 800px){.cv-header__desc{margin-top:2rem}}.cv-header__right{display:flex;border:2px solid purple;border-radius:50%}@media(max-width: 600px){.cv-header__right{margin-top:2rem}}.cv-header__photocon{display:flex;min-width:18rem;height:18rem;padding:2px}.cv-header__photocon-pre{display:contents;padding:2px}.cv-header__photo{border-radius:50%}@media(max-width: 800px){.cv-header__left{display:flex;flex-wrap:wrap;justify-content:center}}.cv-header__nav{width:100%;margin-top:4rem;padding:0 1rem;background-color:rgba(185,161,20,.8)}.cv-header__navblock{justify-content:space-around;margin:0;font:normal 400 1.8rem/1.56 sans-serif}.cv-header__navblock li:not(:last-child){margin-right:1rem}@media(max-width: 600px){.cv-header__navblock li:not(:last-child){width:100%;border-bottom:#000 solid 2px}}.cv-header__navblock>li{position:relative;display:block;text-align:center}.cv-header__navblock>li>a{position:relative;display:inline-block;padding:2rem}.cv-header__navblock>li>a:before{content:"";position:absolute;width:100%;height:3px;top:0;left:0;background-color:#000;transform:scaleX(0);transform-origin:top left;transition:transform .3s}.cv-header__navblock>li>a:after{content:"";position:absolute;width:100%;height:3px;bottom:0;right:0;background-color:#000;transform:scaleX(0);transform-origin:bottom right;transition:transform .3s}.cv-header__navblock>li>a:hover:before{transform-origin:top right;transform:scaleX(1)}.cv-header__navblock>li>a:hover:after{transform-origin:bottom left;transform:scaleX(1)}.cv-main{justify-content:flex-start;align-items:flex-start}.cv-main__section{margin-top:1rem;padding-top:1rem;border-top:2px solid rgba(185,161,20,.8)}.cv-main__section-header{margin-top:3rem;font:normal 400 2rem/1.4 sans-serif;color:#73808d}.cv-main__section-string{justify-content:space-between;margin-bottom:2rem}.cv-main__left{flex-direction:column;flex-wrap:wrap;width:65%;font:normal 400 1.8rem/1.4 sans-serif;padding-right:2rem;border-right:rgba(185,161,20,.8) solid 2px}@media(max-width: 1100px){.cv-main__left{width:100%;padding-right:0;border:none}}.cv-main__block{margin-top:3rem}.cv-main__block-date{font:normal 400 1.4rem/2 sans-serif;color:#73808d}.cv-main__about-header,.cv-main__block-header{font:normal 700 2.2rem/1.5 sans-serif}.cv-main__right{color:#73808d;font:normal 400 2.2rem/1.4 sans-serif}@media(max-width: 1100px){.cv-main__right{display:flex;flex-direction:row;flex-wrap:wrap;margin-left:0;align-items:baseline;width:90%;justify-content:space-between}}.cv-main__right-title{font:normal 700 2.2rem/1.4 sans-serif}.cv-main__right-block{width:100%}.cv-main__right-block:nth-child(2)>.cv-main__right-graph{flex-wrap:wrap;justify-content:center}@media(max-width: 1100px){.cv-main__right-block:nth-child(2)>.cv-main__right-graph{display:flex}}.cv-main__right-progressbar{position:relative;width:17rem;height:17rem;margin:1em;transform:rotate(-90deg)}.cv-main__right-progressbar-svg{position:relative;width:100%;height:100%}.cv-main__right-progressbar-svg-circle{width:100%;height:100%;fill:none;stroke-width:10;stroke-dasharray:440;stroke-dashoffset:440;stroke:#73808d;stroke-linecap:round;transform:translate(5px, 5px)}.cv-main__right-progressbar-text{position:absolute;top:50%;left:50%;padding:.25em .5em;color:#000;border-radius:.25em;font-size:1.8rem;transform:translate(-50%, -50%) rotate(90deg)}.cv-footer{margin-top:3rem;background-color:rgba(185,161,20,.8)}.cv-footer__block{width:100%;padding:2rem 4.8rem;justify-content:space-between;font:normal 400 1.8rem/1.4 sans-serif;min-height:8rem}@media(max-width: 1100px){.cv-footer__block{flex-direction:column}}.cv-footer__section{display:flex;flex-direction:row;justify-content:space-between;align-items:center}.cv-footer__logo{width:7rem;height:4rem}.cv-footer__git,.cv-footer__year{align-self:center}.return_button{background-color:rgba(185,161,20,.8);opacity:0;padding:1.5rem;border-radius:50%;position:fixed;right:4rem;bottom:9rem;cursor:pointer;transition:all .3s ease-in-out}.m-top{margin-top:2rem}.m-left{margin-left:2rem}.shadow-html,.shadow-css,.shadow-js,.shadow-git,.shadow-github,.shadow-webpack,.shadow-sass{filter:drop-shadow(0 0 5px rgba(185, 161, 20, 0.8))}.circle-html{-webkit-animation:anim_circle-html 1s ease-in-out forwards;animation:anim_circle-html 1s ease-in-out forwards}.circle-css{-webkit-animation:anim_circle-css 1s ease-in-out forwards;animation:anim_circle-css 1s ease-in-out forwards}.circle-js{-webkit-animation:anim_circle-js 1s ease-in-out forwards;animation:anim_circle-js 1s ease-in-out forwards}.circle-git{-webkit-animation:anim_circle-git 1s ease-in-out forwards;animation:anim_circle-git 1s ease-in-out forwards}.circle-github{-webkit-animation:anim_circle-github 1s ease-in-out forwards;animation:anim_circle-github 1s ease-in-out forwards}.circle-sass{-webkit-animation:anim_circle-github 1s ease-in-out forwards;animation:anim_circle-github 1s ease-in-out forwards}.circle-webpack{-webkit-animation:anim_circle-webpack 1s ease-in-out forwards;animation:anim_circle-webpack 1s ease-in-out forwards}@-webkit-keyframes anim_circle-html{to{stroke-dashoffset:0}}@keyframes anim_circle-html{to{stroke-dashoffset:0}}@-webkit-keyframes anim_circle-css{to{stroke-dashoffset:50}}@keyframes anim_circle-css{to{stroke-dashoffset:50}}@-webkit-keyframes anim_circle-js{to{stroke-dashoffset:110}}@keyframes anim_circle-js{to{stroke-dashoffset:110}}@-webkit-keyframes anim_circle-git{to{stroke-dashoffset:50}}@keyframes anim_circle-git{to{stroke-dashoffset:50}}@-webkit-keyframes anim_circle-github{to{stroke-dashoffset:0}}@keyframes anim_circle-github{to{stroke-dashoffset:0}}@-webkit-keyframes anim_circle-sass{to{stroke-dashoffset:50}}@keyframes anim_circle-sass{to{stroke-dashoffset:50}}@-webkit-keyframes anim_circle-webpack{to{stroke-dashoffset:400}}@keyframes anim_circle-webpack{to{stroke-dashoffset:400}}@media(max-width: 800px){.cv-footer__git{text-align:center}.cv-footer__section{justify-content:space-evenly;order:1}.cv-footer__year{order:2}}
